+++ src/registers.rs
@@ -15,15 +15,17 @@ pub const FDEV_LSB: u8 = 0x05;
 // RegRxConfig -------------------------------------------------------------------------------------
 pub const RX_CONFIG: u8 = 0x0d;
 pub const RX_CONFIG_RESTART_RX_ON_COLLISION_MASK: u8 = 0x80;
-pub const RX_CONFIG_RESTART_WITHOUT_PLL_LOCK_MASK: u8 = 0x40;
+pub const RX_CONFIG_RESTART_WITHOUT_PLL_LOCK_MASK: u8 = 0x40; // TODO impl
 
-pub const RX_CONFIG_RESTART_WITH_PLL_LOCK_MASK: u8 = 0x20;
+pub const RX_CONFIG_RESTART_WITH_PLL_LOCK_MASK: u8 = 0x20; // TODO impl
 pub const RX_CONFIG_AFC_AUTO_ON_MASK: u8 = 0x10;
 pub const RX_CONFIG_AGC_AUTO_ON_MASK: u8 = 0x08;
 pub const RX_CONFIG_RX_TRIGGER_MASK: u8 = 0x07;
 
-
-
+// RegRssiConfig -------------------------------------------------------------------------------------
+pub const RSSI_CONFIG: u8 = 0x0e;
+pub const RSSI_CONFIG_RSSI_OFFSET_MASK: u8 = 0xf8; // TODO impl
+pub const RSSI_CONFIG_RSSI_SMOOTHING_MASK: u8 = 0x07;
 
 // RegRxBw -----------------------------------------------------------------------------------------
 pub const RX_BW: u8 = 0x12;