+++ CHANGELOG.md
@@ -12,3 +12,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `bit_rate`, `set_bit_rate` methods
 - `fdev`, `set_fdev` methods
 - `set_frequency`, `set_rx_config` methods
+- `RssiSmoothing` enum and `set_rssi_smoothing` method